talk-to-docs/chat/requirements.txt
langchain==0.0.354
streamlit==1.29.0
google-cloud-aiplatform==1.38.1
google-cloud-logging==3.9.0
google-cloud-discoveryengine
cloud-sql-python-connector
psycopg2-binary
pgvector==0.1.8
asyncpg==0.27.0 
pymysql
pg8000
